WIP
got things to work for ns queries
and cluster and node, and podlogs
it looks like more code right now but its just bc of the large
amount of copy/paste and commenting out of code instead of
deleting it.

seems like i should just blow stuff up and rewrite it entirely
but that makes most sense when we can make it a plugin.

Can we make it a plugin that we run after the others?